rename atrium-scala to atrium-scala2 and change package structure
Don't prefix packages with scala but move it after ch.tutteli.atrium.
IMO it is not an extension like kotlin_1_3 and therefore should
not be at the end. Use scala2 as we are likely going to use atrium-scala
for scala 3 once it is out.
