+++ .github/workflows/java-windows.yml
@@ -47,9 +47,9 @@ jobs:
                 run: ATRIUM_ANDROID_JAR="$PWD/android-jar-cache/android.jar" ./gradlew checkDexer
                 shell: bash
 
-            -   name: composite build atrium-scala
+            -   name: composite build atrium-scala2
                 run: ./gradlew build
-                working-directory: misc\tools\atrium-scala-test
+                working-directory: misc\tools\atrium-scala2-test
 
             -   name: Upload windows build code coverage
                 uses: codecov/codecov-action@v1
