+++ .travis.yml
@@ -13,7 +13,7 @@ jobs:
             - set -e
             - ./gradle/scripts/check-generateLogic-committed.sh
             - ./gradlew buildAllWithoutJs
-            - misc/tools/atrium-scala-test/gradlew -p ./misc/tools/atrium-scala-test build
+            - misc/tools/atrium-scala2-test/gradlew -p ./misc/tools/atrium-scala2-test build
         after_success: bash <(curl -s https://codecov.io/bash) -F current
 
       - env: MODE=current_8
