--- atrium-scala/apis/atrium-scala-api-fluent-en_GB/src/main/scala/ch/tutteli/atrium/api/fluent/en_GB/scala/BaseAssertions.scala
@@ -1,9 +1,0 @@
-package ch.tutteli.atrium.api.fluent.en_GB.scala
-
-import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.domain.creating.changers.PostFinalStep
-trait BaseAssertions {
-  implicit class RxPostFinalStep[T, R, E <: Expect[R]](postStep: PostFinalStep[T, R, E]) {
-    @inline def expectOfFeature: E = postStep.getExpectOfFeature
-  }
-}
